% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{colour_palette_CREG}
\alias{colour_palette_CREG}
\title{Colour palette}
\usage{
colour_palette_CREG(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{A string (possibilities are: "countries", "main", "blue")}

\item{reverse}{Logical, define whether the colours need to be applied in reverse order (default = FALSE)}

\item{...}{other}
}
\description{
This function defines colour palettes for ggplot objects.
}
