% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{CREG_palette}
\alias{CREG_palette}
\title{CREG palette with ramped colours}
\usage{
CREG_palette(palette = "default", alpha = 1, reverse = FALSE)
}
\arguments{
\item{palette}{}

\item{alpha}{}

\item{reverse}{}
}
\value{

}
\description{
CREG palette with ramped colours
}
