% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left join.R
\name{left_join_CREG}
\alias{left_join_CREG}
\title{Left join csv files which were extracted from DWH}
\usage{
left_join_CREG(subfolder, filename, col_types)
}
\arguments{
\item{filename}{A stringn, including the name of hte csv file (don't include the extension ".csv" as this is added automatically)}

\item{col_types}{A string describing, in short form, the column types of the data itself (i.e. not the year, month, day, etc. variables as those are added automatically)}

\item{subfoler}{A string, including the subfolder where the csv file is stored}
}
\description{
This function left joints a csv file to another csv.
The arguments include a reference to a subfolder of the working directory(filepath), a reference to the csv file (filename),
and an indication of the type of column (in abbreviate form, e.g. "n" = numeric, "d" = double, "c" = character).
}
\examples{
df <- loadcsv_CREG(filepath = "source data", filename = "loadBE", col_types = "n") \%>\%
left_join(filepath = "source data", filename = "loadNL", col_types = "n")
}
