% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load csv.R
\name{loadmultiplecsv_CREG}
\alias{loadmultiplecsv_CREG}
\title{Import multiple csv files which were extracted from DWH and join them}
\usage{
loadmultiplecsv_CREG(subfolder, filenames, col_types, resolution)
}
\arguments{
\item{col_types}{A string vector, describing, in short form, the column types of the data itself (i.e. not the year, month, day, etc. variables as those are added automatically) (e.g. c("nn", "nnnn"))}

\item{resolution}{A string vector, describing whether hourly ("h") or quarterly ("q") data are inmported (e.g. "q")}

\item{subfodler}{A string, including the subfolder where the csv file is stored}

\item{filename}{A string vector, including the name of the csv file (don't include the extension ".csv" as this is added automatically) (e.g. c("loadBE", "loadNL"))}
}
\description{
This function imports a csv file into an R session.
The arguments include a reference to a subfolder of the working directory(filepath), a reference to the csv files (filenames),
and an indication of the type of column (in abbreviate form, e.g. "n" = numeric, "d" = double, "c" = character).
}
\examples{
df <- loadmultiplecsv_CREG(filepath = "source data", filename = c("loadBE", "loadNL"), col_types = c("n", "n"), quarterlyresolution = c("h", "h"))
}
