% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save ggplot.R
\name{save_CREG}
\alias{save_CREG}
\title{Save ggplot to png file to use in report}
\usage{
save_CREG(filename)
}
\arguments{
\item{filename}{A string}
}
\description{
This function saves the last plot that was displayed, in a png file with the dimensions (16 x 10 cm) for easy processing in Word reports. The document is saved, by default, in the working directory.
}
\examples{
save_CREG(filename = "figure")
}
