% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_CREG}
\alias{scale_colour_CREG}
\title{CREG colour palette for ggplot2}
\usage{
scale_colour_CREG(
  ...,
  palette = "default",
  discrete = TRUE,
  alpha = 1,
  reverse = FALSE
)
}
\arguments{
\item{...}{}

\item{palette}{}

\item{discrete}{}

\item{alpha}{}

\item{reverse}{}
}
\value{

}
\description{
CREG colour palette for ggplot2
}
