% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_dutchmasters}
\alias{scale_fill_dutchmasters}
\title{CREG fill palette for ggplot2}
\usage{
scale_fill_dutchmasters(
  ...,
  palette = "milkmaid",
  discrete = TRUE,
  alpha = 1,
  reverse = FALSE
)
}
\arguments{
\item{...}{}

\item{palette}{}

\item{discrete}{}

\item{alpha}{}

\item{reverse}{}
}
\value{

}
\description{
CREG fill palette for ggplot2
}
