% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_CREG}
\alias{theme_CREG}
\title{Apply custom CREG theme to ggplot visuals}
\usage{
theme_CREG()
}
\value{

}
\description{
This function applies a custom theme, based on the CREG Style Guide, for the lay-out and colors of ggplot visuals. The colors are based on the CREG pallete, defined in the colours_CREG() function.
}
\examples{
ggplot(data = mpg) +
geom_point(mapping = aes(x = displ, y = hwy)) +
theme_CREG()
}
