% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load csv.R
\name{loadcsv_CREG}
\alias{loadcsv_CREG}
\title{Import csv files which were extracted from DWH}
\usage{
loadcsv_CREG(filepath)
}
\arguments{
\item{filepath}{A string, including (if relevant) the subfolder where the csv file is stored (relative to the current working directory)}
}
\description{
This function imports a csv file into an R session.
The arguments include a reference to a subfolder of the working directory(filepath), a reference to the csv file (filename),
and an indication of the type of column (in abbreviate form, e.g. "n" = numeric, "d" = double, "c" = character).
}
\examples{
df <- loadcsv_CREG(filepath = "source data/loadBelgium.csv")
}
